Add support for interpolated strings.
Interpolated strings have syntax and semantics similar to PEP 498's
formatted string literals. The syntax accepts more kinds of expressions
as replacements than PEP 498, but does not allow the `!a` conversion
specifier or the use of replacements as format specifiers. Interpolated
strings compile to a call to `string.format`. The receiver is the
literal parts of the interpolated string separated by replacement fields
that contain appropriate conversions and format specifiers; the
arguments are the expression parts of the interpolated string.

This is related to https://github.com/bazelbuild/starlark/issues/91, and
is a proof point for a simple implementation